==> 01_dowloading_data_with_allensdk_tests.ipynb <==
# this notebook is just testing how to download data, what data is available, etc.

# !python -m pip install --upgrade pip
# !pip install mindscope_utilities --upgrade
import pandas as pd
import numpy as np
from tqdm import tqdm

from allensdk.brain_observatory.behavior.behavior_project_cache import VisualBehaviorOphysProjectCache
import mindscope_utilities.visual_behavior_ophys as ophys

cache = VisualBehaviorOphysProjectCache.from_s3_cache(cache_dir='data')
# session_table = cache.get_ophys_session_table()
experiment_table = cache.get_ophys_experiment_table()
experiment_table = experiment_table.reset_index()

# from the experiment table we only want those where project_code is VisualBehaviorMultiscope
multiscope_experiment_ids = experiment_table[experiment_table['project_code'] == 'VisualBehaviorMultiscope']['ophys_experiment_id'].tolist()

# reminder: an experiment is a single imaging plane acquired in a single session
# https://allensdk.readthedocs.io/en/latest/visual_behavior_optical_physiology.html

len(multiscope_experiment_ids) # how many multiscope experiments we have

# for now let's try only 5
multiscope_experiment_ids = multiscope_experiment_ids[:5]

# this loop downloads behavioural data for each of the experiments whose ids we found earlier and saves them in a dictionary where the key is experiment id
experiments = {}
for id in multiscope_experiment_ids:
  experiments[id] = cache.get_behavior_ophys_experiment(id)

experiments[951980471]
# each experiment is a BehaviorOpysExperiment object
# there is a table at the bottom of this page with description of what such objects contain and how to access that information
# https://allensdk.readthedocs.io/en/v2.12.3/visual_behavior_ophys_nwb.html

# for example this shows rewards data for experiment 951980471
experiments[951980471].rewards

# this shows dff data for experiment 951980471
experiments[951980471].dff_traces.reset_index()

# metadata available for this experiment (this will be useful for downloading neural data later)
experiments[951980471].metadata

# list of metadata we want to download with out neural data
metadata_keys = ['cre_line', 'imaging_depth', 'mouse_id', 'ophys_experiment_id', 'session_type', 'targeted_structure']

# this downloads the neural data (copied from load_Allen_Visual_Behavior_from_SDK)
neural_data = []
for ophys_experiment_id in tqdm(experiments.keys()):
  this_experiment = experiments[ophys_experiment_id]
  this_experiment_neural_data = ophys.build_tidy_cell_df(this_experiment)

  for metadata_key in metadata_keys:
    this_experiment_neural_data[metadata_key] = this_experiment.metadata[metadata_key]

  # append the data for this experiment to a list
  neural_data.append(this_experiment_neural_data)

# concatate the list of dataframes into a single dataframe
neural_data = pd.concat(neural_data)

neural_data # once we have neural data in a pandas dataframe like this it's much easier to work with it

neural_data[neural_data['ophys_experiment_id'] == 951980471] # neural data for one experiment

stimulus_presentations_one_experiment = experiments[951980471].stimulus_presentations.reset_index().drop(columns = ['image_set'])
stimulus_presentations_one_experiment.insert(0, 'ophys_experiment_id', [951980471 for i in range(stimulus_presentations_one_experiment.shape[0])]) # adds column with experiment id
stimulus_presentations_one_experiment

experiments[951980471].eye_tracking.reset_index() # eyes data from one experiment but from this we probably only need pupil_area

experiments[951980471].trials.reset_index().columns 
# this one has information about rewards (time and volume) also the type of trial (hit, miss, false alarm, correct rejection)
# don't know if we need it but we know it's available too

# So in general we can actually get all the data in nice data frames but it will require some processing (joining, etc.)
