==> 03_random_cell_etr.ipynb <==
# this notebook is for analysing data after downloading it from google drive to your computer

import os

import pandas as pd
import numpy as np
from icecream import ic
from tqdm import tqdm

from event_triggered_response import *

import seaborn as sns
sns.set_theme(style="darkgrid")
sns.set_palette('colorblind')

all_data_folders = os.listdir('data')

multiscope_experiment_table = pd.read_csv('multiscope_experiment_table_filtered.csv', index_col='ophys_experiment_id').reset_index().drop(columns = ['Unnamed: 0'])
multiscope_experiment_table.head()

if not os.path.exists('cell_etr_data'):
    os.mkdir('cell_etr_data')

for folder in tqdm(all_data_folders):
    try:
        neural_data = pd.read_csv(f'data/{folder}/neural_data.csv').drop(columns = ['Unnamed: 0'])
        stimulus_presentations = pd.read_csv(f'data/{folder}/stimulus_presentations.csv').drop(columns = ['Unnamed: 0'])
        for id in tqdm(np.random.choice(neural_data['ophys_experiment_id'].unique().tolist(), size = 2, replace = False)):
            n = neural_data[neural_data['ophys_experiment_id'] == id]
            s = stimulus_presentations[stimulus_presentations['ophys_experiment_id'] == id]
            for cell in tqdm(np.random.choice(n['cell_specimen_id'].unique().tolist(), size = 5, replace = False)):
                etr = event_triggered_response(data=n[n['cell_specimen_id'] == cell],
                                               t='timestamps',
                                               y='dff',
                                               event_times=s[s['omitted'] == False]['start_time'],
                                               t_before=3,
                                               t_after=3,
                                               output_sampling_rate=50)
                etr.insert(0, 'ophys_experiment_id', [id for i in range(etr.shape[0])])
                etr.insert(0, 'cell_specimen_id', [cell for i in range(etr.shape[0])])
                etr.to_csv(f'cell_etr_data/{id}_{cell}.csv')
    except:
        print(folder)
