==> 07_event_triggered_response_reshaping_for_classifier_google_colab_drive.ipynb <==
!python -m pip install --upgrade pip --quiet
!pip install mindscope_utilities --upgrade --quiet
!pip install icecream --quiet

import pandas as pd
import numpy as np
from icecream import ic # icecream is an alternative to print useful for debugging code
# https://github.com/gruns/icecream
import os

# this part is for connecting with google drive and saving data there so we don't need to download again
# you might get a prompt asking you to authorize
from google.colab import drive
drive.mount('/content/gdrive')
# https://towardsdatascience.com/downloading-datasets-into-google-drive-via-google-colab-bcb1b30b0166

# copy the data to your own google drive - this whole folder:
# https://drive.google.com/drive/folders/18_Bin6vrtffX5iAim1KxCh1SJkvKYpM0?usp=sharing

etr_filenames = os.listdir('/content/gdrive/MyDrive/neuromatch/mean_cell_etr_data') # replace with path where you copied mean etr cell data from me
# each file in this folder is one random experiment, event triggered responses for all cells in that experiment for a total of 25 experiments and about 125 cells

all_etrs = []
for name in etr_filenames:
  etr = pd.read_csv(f'/content/gdrive/MyDrive/neuromatch/mean_cell_etr_data/{name}').drop(columns = ['Unnamed: 0'])
  all_etrs.append(etr)
all_etrs = pd.concat(all_etrs).reset_index(drop=True)

all_etrs

wide_etrs = pd.pivot(all_etrs[['ophys_experiment_id', 'cell_specimen_id', 'cre_line', 'experience_level', 'time', 'dff']],
         index = ['ophys_experiment_id', 'cell_specimen_id', 'cre_line', 'experience_level'],
         columns = 'time',
         values = 'dff').reset_index().drop(columns = ['ophys_experiment_id', 'cell_specimen_id'])
wide_etrs

# replacing experience level with just novel / familiar:
new_vals = [x.lower() if x == 'Familiar' else 'novel' for x in wide_etrs['experience_level']] # this trick is called list comprehension it is kind of a for loop in one line for lists
wide_etrs['experience_level'] = new_vals
wide_etrs.head(20) # if all are novel something is wrong, try to rerun everything from the beginning

x = wide_etrs.drop(columns = 'experience_level')
y = wide_etrs['experience_level']
x.shape

x = pd.get_dummies(x) # this splits cre line into two columns with 0-1
# it is a very simplified approach to one hot encoding 
# one hot encoding is explained here at the beginning https://www.kaggle.com/code/dansbecker/using-categorical-data-with-one-hot-encoding/notebook
x.head()
